# openapi/control.openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "additionalProperties": false,
        "description": "Stable HTTP error envelope exposed by API-owned routes.",
        "properties": {
          "detail": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CourseCode": {
        "description": "Nominal nullable course identity supplied by an owning adapter.",
        "maxLength": 32,
        "minLength": 2,
        "pattern": "^[A-Z0-9-]{2,32}$",
        "title": "CourseCode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LifeDomain": {
        "enum": [
          "research",
          "teaching",
          "software",
          "business",
          "personal",
          "administration",
          "finance",
          "health",
          "media",
          "other"
        ],
        "title": "LifeDomain",
        "type": "string"
      },
...
      "OrganizerActivity": {
        "enum": [
          "off",
          "paused",
          "idle",
          "discovering",
          "scanning",
          "classifying",
          "failed"
        ],
        "title": "OrganizerActivity",
        "type": "string"
      },
      "OrganizerControlMode": {
        "enum": [
          "on",
          "pause",
          "off"
        ],
        "title": "OrganizerControlMode",
        "type": "string"
      },
      "OrganizerControlRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/OrganizerControlMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "OrganizerControlRequest",
        "type": "object"
      },
      "OrganizerFileAvailability": {
        "enum": [
          "local",
          "cloud_placeholder",
          "inaccessible"
        ],
        "title": "OrganizerFileAvailability",
        "type": "string"
      },
      "OrganizerParaCategory": {
        "enum": [
          "project",
          "area",
          "resource",
          "archive",
          "inbox"
        ],
        "title": "OrganizerParaCategory",
        "type": "string"
      },
      "OrganizerProposalListResponse": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/OrganizerProposalResponse"
            },
            "maxItems": 500,
            "title": "Proposals",
            "type": "array"
          },
          "total": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "proposals",
          "total",
          "complete"
        ],
        "title": "OrganizerProposalListResponse",
        "type": "object"
      },
      "OrganizerProposalResponse": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/OrganizerFileAvailability"
          },
          "byte_size": {
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "course_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CourseCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "extension": {
            "maxLength": 128,
            "pattern": "^(|\\.[^/\\\\\\x00-\\x1f\\x7f]+)$",
            "title": "Extension",
            "type": "string"
          },
          "file_id": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "File Id",
            "type": "string"
          },
          "life_domain": {
            "$ref": "#/components/schemas/LifeDomain"
          },
          "model": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "model_digest": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Model Digest",
            "type": "string"
          },
          "name": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "para_category": {
            "$ref": "#/components/schemas/OrganizerParaCategory"
          },
          "proposed_at": {
            "format": "date-time",
            "title": "Proposed At",
            "type": "string"
          },
          "rationale": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "relative_path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Relative Path",
            "type": "string"
          },
          "root_id": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Root Id",
            "type": "string"
          },
          "suggested_group": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Suggested Group",
            "type": "string"
          }
        },
        "required": [
          "file_id",
          "root_id",
          "relative_path",
          "name",
          "extension",
          "byte_size",
          "availability",
          "para_category",
          "life_domain",
          "course_code",
          "confidence",
          "suggested_group",
          "rationale",
          "model",
          "model_digest",
          "proposed_at"
        ],
        "title": "OrganizerProposalResponse",
        "type": "object"
      },
      "OrganizerStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/OrganizerActivity"
          },
          "current_relative_path": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Relative Path"
          },
          "current_root_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 64,
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Root Id"
          },
          "desired_mode": {
            "$ref": "#/components/schemas/OrganizerControlMode"
          },
          "discovered_roots": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Discovered Roots",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_event_sequence": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Last Event Sequence",
            "type": "integer"
          },
          "local_files": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Local Files",
            "type": "integer"
          },
          "observed_files": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Observed Files",
            "type": "integer"
          },
          "placeholder_files": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Placeholder Files",
            "type": "integer"
          },
          "proposed_files": {
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Proposed Files",
            "type": "integer"
          }
        },
        "required": [
          "desired_mode",
          "activity",
          "discovered_roots",
          "observed_files",
          "local_files",
          "placeholder_files",
          "proposed_files",
          "last_event_sequence"
        ],
        "title": "OrganizerStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Project Koios",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/organizer/control": {
      "put": {
        "operationId": "update_control_organizer_control_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizerControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizerStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter failed unexpectedly."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter returned an invalid projection."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No organizer owner adapter is available."
          }
        },
        "summary": "Update Control",
        "tags": [
          "organizer"
        ]
      }
    },
    "/organizer/proposals": {
      "get": {
        "operationId": "read_proposals_organizer_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "life_domain",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LifeDomain",
              "default": "teaching"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizerProposalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter failed unexpectedly."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter returned an invalid projection."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No organizer owner adapter is available."
          }
        },
        "summary": "Read Proposals",
        "tags": [
          "organizer"
        ]
      }
    },
    "/organizer/status": {
      "get": {
        "operationId": "read_status_organizer_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizerStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter failed unexpectedly."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The organizer owner adapter returned an invalid projection."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No organizer owner adapter is available."
          }
        },
        "summary": "Read Status",
        "tags": [
          "organizer"
        ]
      }
    },
...
  }
}